+++ gdal-sys/src/lib.rs
@@ -3,5 +3,6 @@
 #![allow(non_snake_case)]
 #![allow(clippy::upper_case_acronyms)]
 #![allow(rustdoc::bare_urls)]
+
 include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 include!(concat!(env!("OUT_DIR"), "/docs_rs_helper.rs"));